fix(B-464): split confidence score schema and model
