+++ backend/app/routers/user_profile_route.py
@@ -7,7 +7,7 @@ from sqlmodel import select
 
 from app.database import get_db_session
 from app.dependencies import require_user
-from app.models.user_confidence_score import ConfidenceScoreRead, UserConfidenceScore
+from app.models.user_confidence_score import UserConfidenceScore
 from app.models.user_goal import Goal, UserGoal
 from app.models.user_profile import (
     UserProfile,
@@ -16,6 +16,7 @@ from app.models.user_profile import (
     UserProfileReplace,
     UserProfileUpdate,
 )
+from app.schemas.user_confidence_score import ConfidenceScoreRead
 
 router = APIRouter(prefix='/user-profile', tags=['User Profiles'])
